chore: bump to 1.6.3

+++ src/sink/send.rs
@@ -29,7 +29,7 @@ pub fn send_thread(token: &str,
                    todo: Arc<Mutex<VecDeque<PathBuf>>>,
                    timeout: u64,
                    batch_count: u64,
-                   batch_size: i64,
+                   batch_size: u64,
                    sigint: Arc<AtomicBool>) {
     loop {
         Data::new(sigint.clone(), todo.clone(), batch_count, batch_size)
@@ -111,12 +111,12 @@ impl Data {
     pub fn new(sigint: Arc<AtomicBool>,
                todo: Arc<Mutex<VecDeque<PathBuf>>>,
                batch_count: u64,
-               batch_size: i64)
+               batch_size: u64)
                -> Option<Data> {
         let mut d = Data {
             sigint: sigint,
             batch_count: batch_count,
-            batch_size: batch_size,
+            batch_size: batch_size as i64,
             todo: todo,
             processing: None,
             processed: Vec::new(),
